requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
slack_sdk
python-dotenv
